Changes to the IMU code

+++ auto_rover_dfrobot_gravity_10dof/include/auto_rover_dfrobot_gravity_10dof/auto_rover_dfrobot_gravity_10dof.h
@@ -421,21 +421,21 @@ namespace auto_rover_dfrobot_gravity_10dof
 
 		uint32_t CompensateTemperature(int32_t& adc);
 
-		void ReadAccelData(vector<int16_t>& destination);
+		void ReadAccelData(vector<double>& destination);
 
-		void ReadGyroData(vector<int16_t>& destination);
+		void ReadGyroData(vector<double>& destination);
 
 		int8_t ReadGyroTempData();
 
-		void ReadMagData(vector<int16_t>& destination);
+		void ReadMagData(vector<double>& destination);
 
-		void ReadQuatData(vector<int16_t>& destination);
+		void ReadQuatData(vector<double>& destination);
 
-		void ReadEulData(vector<int16_t>& destination);
+		void ReadEulData(vector<double>& destination);
 
-		void ReadLIAData(vector<int16_t>& destination);
+		void ReadLIAData(vector<double>& destination);
 
-		void ReadGRVData(vector<int16_t>& destination);
+		void ReadGRVData(vector<double>& destination);
 
 	private:
 		I2C i2c_;
